+++ webapp/src/components/styles/tailwind.css
@@ -1,0 +1,5 @@
+@import "tailwindcss";
+
+
+/* Not sure if we should have one css file or multiple tbh */
+
