+++ webapp/src/index.css
@@ -1,3 +1,6 @@
+
+/* Default CSS */
+
 :root {
   font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
   line-height: 1.5;
@@ -66,3 +69,6 @@ button:focus-visible {
     background-color: #f9f9f9;
   }
 }
+
+
+
